+++ src/lib.rs
@@ -52,6 +52,7 @@
 //! ```
 
 #![cfg_attr(not(any(test, feature = "fakes")), no_std)]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 #![deny(clippy::undocumented_unsafe_blocks)]
 #![deny(unsafe_op_in_unsafe_fn)]